Add visual polish and UI effects to menus and screens
- Add gradient and panel drawing methods to Renderer
- Implement animated particle background for main menu
- Polish virtual keyboard with gradient keys and glow effects
- Add gradient backgrounds and styled panels to all menu screens
- Polish game setup screen with styled player slots
- Add animated stage selection cards with pulsing borders
- Fix CMake shared/static library configuration for SDL3_ttf
- Fix narrowing conversion warnings in SDL_Color initialization

+++ src/battle.cpp
@@ -1128,7 +1128,10 @@ void BattleManager::renderGameOver(const BattleState& state) {
                              renderer.getFontSmall(), {255, 255, 255, 255}, TextAlign::Center);
         } else {
             // Not confirmed - dim outline
-            SDL_Color dimColor = {playerColor.r / 2, playerColor.g / 2, playerColor.b / 2, 150};
+            SDL_Color dimColor = {
+                static_cast<Uint8>(playerColor.r / 2),
+                static_cast<Uint8>(playerColor.g / 2),
+                static_cast<Uint8>(playerColor.b / 2), 150};
             renderer.drawRectOutline(indicatorX, confirmY, 30, 30, dimColor, 2.0f);
             renderer.drawText("?", indicatorX + 15, confirmY + 6,
                              renderer.getFontSmall(), dimColor, TextAlign::Center);
